azalea_core::position
Struct ChunkSectionBlockPos

pub struct ChunkSectionBlockPos {
    pub x: u8,
    pub y: u8,
    pub z: u8,
}

The coordinates of a block inside a chunk section. Each coordinate must be in the range [0, 15].

Fields
x: u8y: u8z: u8
Implementations

impl ChunkSectionBlockPos

pub fn new(x: u8, y: u8, z: u8) -> Self

pub fn length_sqr(&self) -> u8

Get the distance of this vector to the origin by doing x^2 + y^2 + z^2.

pub fn distance_to_sqr(&self, other: &Self) -> u8

Get the squared distance from this position to another position. Equivalent to (self - other).length_sqr().

pub fn horizontal_distance_sqr(&self) -> u8

pub fn horizontal_distance_to_sqr(&self, other: &Self) -> u8

pub fn down(&self, y: u8) -> Self

Return a new instance of this position with the y coordinate decreased by the given number.

pub fn up(&self, y: u8) -> Self

Return a new instance of this position with the y coordinate increased by the given number.

pub fn north(&self, z: u8) -> Self

Return a new instance of this position with the z coordinate subtracted by the given number.

pub fn east(&self, x: u8) -> Self

Return a new instance of this position with the x coordinate increased by the given number.

pub fn south(&self, z: u8) -> Self

Return a new instance of this position with the z coordinate increased by the given number.

pub fn west(&self, x: u8) -> Self

Return a new instance of this position with the x coordinate subtracted by the given number.

pub fn dot(&self, other: Self) -> u8