azalea_core::position
Struct ChunkSectionPos

pub struct ChunkSectionPos {
    pub x: i32,
    pub y: i32,
    pub z: i32,
}

The coordinates of a chunk section in the world.

Fields
x: i32y: i32z: i32
Implementations

impl ChunkSectionPos

pub fn new(x: i32, y: i32, z: i32) -> Self

pub fn length_sqr(&self) -> i32

Get the distance of this vector to the origin by doing x^2 + y^2 + z^2.

pub fn distance_to_sqr(&self, other: &Self) -> i32

Get the squared distance from this position to another position. Equivalent to (self - other).length_sqr().

pub fn horizontal_distance_sqr(&self) -> i32

pub fn horizontal_distance_to_sqr(&self, other: &Self) -> i32

pub fn down(&self, y: i32) -> Self

Return a new instance of this position with the y coordinate decreased by the given number.

pub fn up(&self, y: i32) -> Self

Return a new instance of this position with the y coordinate increased by the given number.

pub fn north(&self, z: i32) -> Self

Return a new instance of this position with the z coordinate subtracted by the given number.

pub fn east(&self, x: i32) -> Self

Return a new instance of this position with the x coordinate increased by the given number.

pub fn south(&self, z: i32) -> Self

Return a new instance of this position with the z coordinate increased by the given number.

pub fn west(&self, x: i32) -> Self

Return a new instance of this position with the x coordinate subtracted by the given number.

pub fn dot(&self, other: Self) -> i32

impl ChunkSectionPos

pub fn block_to_section_coord(block: i32) -> i32