azalea_core::position
Struct Vec3

pub struct Vec3 {
    pub x: f64,
    pub y: f64,
    pub z: f64,
}

Used to represent an exact position in the world where an entity could be. For blocks, BlockPos is used instead.

Fields
x: f64y: f64z: f64
Implementations

impl Vec3

pub fn with_delta(&self, delta: &impl PositionDeltaTrait) -> Vec3

pub fn length_squared(&self) -> f64

pub fn normalize(&self) -> Vec3

pub fn multiply(&self, x: f64, y: f64, z: f64) -> Vec3

pub fn scale(&self, amount: f64) -> Vec3

impl Vec3

pub fn new(x: f64, y: f64, z: f64) -> Self

pub fn length_sqr(&self) -> f64

Get the distance of this vector to the origin by doing x^2 + y^2 + z^2.

pub fn distance_to_sqr(&self, other: &Self) -> f64

Get the squared distance from this position to another position. Equivalent to (self - other).length_sqr().

pub fn horizontal_distance_sqr(&self) -> f64

pub fn horizontal_distance_to_sqr(&self, other: &Self) -> f64

pub fn down(&self, y: f64) -> Self

Return a new instance of this position with the y coordinate decreased by the given number.

pub fn up(&self, y: f64) -> Self

Return a new instance of this position with the y coordinate increased by the given number.

pub fn north(&self, z: f64) -> Self

Return a new instance of this position with the z coordinate subtracted by the given number.

pub fn east(&self, x: f64) -> Self

Return a new instance of this position with the x coordinate increased by the given number.

pub fn south(&self, z: f64) -> Self

Return a new instance of this position with the z coordinate increased by the given number.

pub fn west(&self, x: f64) -> Self

Return a new instance of this position with the x coordinate subtracted by the given number.

pub fn dot(&self, other: Self) -> f64

impl Vec3

pub fn length(&self) -> f64

Get the distance of this vector to the origin by doing sqrt(x^2 + y^2 + z^2).

pub fn distance_to(&self, other: &Self) -> f64

Get the distance from this position to another position. Equivalent to (self - other).length().