azalea_core
Module registry_holder

The data sent to the client in the ClientboundRegistryDataPacket.

This module contains the structures used to represent the registry sent to the client upon login. This contains a lot of information about the game, including the types of chat messages, dimensions, and biomes.

Structs

AdditionsSound

BiomeEffects

The effects of a biome.

BiomeMoodSound

BiomeMusic

The music of the biome.

BiomeParticle

Biome particles.

ChatTypeData

Data about a chat message.

ChatTypeElement

Data about a kind of chat message

ChatTypeStyle

The style of a chat message.

DamageTypeElement

DimensionTypeElement

Dimension attributes.

MonsterSpawnLightLevelValues

The min and max light levels at which monsters can spawn.

RegistryHolder

The base of the registry.

RegistryType

A collection of values for a certain type of registry data.

TrimMaterialElement

TrimPatternElement

WorldTypeElement

Biome attributes.

Enums

BiomePrecipitation

The precipitation of a biome.

MonsterSpawnLightLevel

The light level at which monsters can spawn.